fixed deletePost

+++ src/services/post.service.ts
@@ -7,13 +7,14 @@ import {
   PostRepository,
   CommentRepository,
   FriendRepository,
-  UserRepository,
+  VoteRepository,
 } from '../repositories';
 import {PolkadotJs} from '../utils/polkadotJs-utils';
-import {injectable, BindingScope} from '@loopback/core';
+import {injectable, BindingScope, service} from '@loopback/core';
 import {BcryptHasher} from './authentication/hash.password.service';
 import {config} from '../config';
 import {PlatformType} from '../enums';
+import {MetricService} from '../services';
 
 @injectable({scope: BindingScope.TRANSIENT})
 export class PostService {
@@ -26,8 +27,10 @@ export class PostService {
     protected commentRepository: CommentRepository,
     @repository(FriendRepository)
     protected friendRepository: FriendRepository,
-    @repository(UserRepository)
-    protected userRepository: UserRepository,
+    @repository(VoteRepository)
+    protected voteRepository: VoteRepository,
+    @service(MetricService)
+    protected metricService: MetricService,
   ) {}
 
   async createPost(post: Omit<ExtendedPost, 'id'>): Promise<PostWithRelations> {
@@ -67,11 +70,21 @@ export class PostService {
     return Object.assign(newPost, {people: people});
   }
 
-  async deletePost(id: string): Promise<void> {
+  async deletePost(id: string, userId?: string): Promise<void> {
+    if (!userId)
+      throw new HttpErrors.UnprocessableEntity('UserId must be filled');
+
+    const {createdBy} = await this.postRepository.findById(id);
+
+    if (createdBy !== userId)
+      throw new HttpErrors.UnprocessableEntity('Only post owner can delete');
+
     await this.postRepository.deleteById(id);
     await this.commentRepository.deleteAll({
       postId: id,
     });
+    await this.voteRepository.deleteAll({postId: id});
+    await this.metricService.userMetric(createdBy);
   }
 
   async getPostImporterInfo(
